_macros.html: Graphs a bit smaller, FlatUI colors.
Currently the graphs are shown at the top of the page and because of
that really take up a lot of screen real estate. This commit makes the
graphs a bit smaller but we really need to rethink the layout here.

Additionally we switch the colors to the FlatUI swatches to match the
rest of the theme better.