puppetboard/app Allowing administrators to configure access to catalog
Adding new configuration option in default_settings.py with the name
ENABLE_CATALOG. Default value is true.

Adding supportin app.py for this new config option. If true render the
template as before. If false return a 403 to the user. Much like the
ENABLE_QUERY option works for the /query interface is.