puppetboard/default_settings.py: Reverting PUPPETDB_HOST change from #187
This change is only useful for Docker image building, but for officially
supported packages in PIP this change will cause unnecessary breakage to
existing users. We will have to implement a different approach for configuring
settings in Docker images.

Adding a note to README.rst indicating that an officially supported Docker
image is planned.