Fix: default table sort was always ascending
An ascending sort order is always triggered when then target index is different
from the tablesort index. This was always true on a page (re)load, because
index is initialized with null.  By setting the instance' index to the
default-sort column index we assure that the direction parameter to sort() is
used.

See https://github.com/kylefox/jquery-tablesort/blob/master/jquery.tablesort.js#L42