Allow to configure enabled query endpoints
Allow to configure which PuppetDB endpoints are allowed in the /query
page, adding the ENABLED_QUERY_ENDPOINTS configuration variable, that
if set restricts the allowed endpoints to query. If not set the previous
behaviour is kept, allowing all PuppetDB endpoints to be queried, if the
ENABLE_QUERY configuration is True.

Closes: #475