Ignore facts environment for compatibility and performance
Because of https://tickets.puppetlabs.com/browse/PUP-9290 if someone
is using `puppet facts upload` their nodes have facts_environment set
always to "production". This breaks filtering in Puppetboard.
But it's both safe and beneficial to just ignore facts_environment
and depend only on catalog_environment to filter nodes by environment
because if a node is in env X then it mean that it has catalog_environment
set to X.