Format complex fact values as pretty, colored JSON
and extend the width of the fact values column in the node
view as the values are usually longer that the names.
Thus this implements #83.

Also fix getting nodes with non-string fact values (#612):
Examples:
* int (in this case the filtering did not occur at all
  and ALL results were returned),
* list of dict (here the filtering was not done and
  ZERO results were returned),

As showing fact values does not use the code that strips
trailing 'u' anymore, this also fixes #567.