Added template to display the class view
It displays a table with the list of nodes having at least one event
on a resource declared in a given class.
For each node, it also displays the status of the node as well as the
status of the class (aka the status of the resources that are declared
in this class).