Added template to display the table in the Classes view
Classes view displays a list of classes that had at least one event on a declared resource.
For each class, we have the number of nodes having executed this class (with at least one change on a resource declared in the class), the number of events by status ('failure', 'success', 'noop').