Added a new view "Classes"
This view displays the classes having had at least one event on one of their declared resources. We have to parse the last report of each node to group the events by class.
For each class, we display the number of nodes having executed this class as well as the number of events by status ('failure', 'success' and 'noop').

When we click on a class we display the class view which allows to list the nodes having executed this class (with at least one status change of a resource of this class).